Added cpp program to find XOR from 1 to N and sorted files
